Move test for mimeparser from stress.rs into src/*
